[all][arm] Add support for arm

+++ libs/libcxx/std/Dbg.h
@@ -10,7 +10,7 @@ public:
     friend const Dbg& operator<<(const Dbg& os, const String& value);
     friend const Dbg& operator<<(const Dbg& os, int value);
     friend const Dbg& operator<<(const Dbg& stream, uint32_t value);
-    friend const Dbg& operator<<(const Dbg& stream, size_t value);
+    friend const Dbg& operator<<(const Dbg& stream, unsigned long value);
 
 private:
     int m_fd { 1 };
@@ -28,7 +28,7 @@ public:
     friend const DbgPid& operator<<(const DbgPid& os, const String& value);
     friend const DbgPid& operator<<(const DbgPid& os, int value);
     friend const DbgPid& operator<<(const DbgPid& stream, uint32_t value);
-    friend const DbgPid& operator<<(const DbgPid& stream, size_t value);
+    friend const DbgPid& operator<<(const DbgPid& stream, unsigned long value);
 
     inline int pid() const { return m_pid; }
